presolvers: one directory each, matching the constraint layout
Presolvers were flat files directly under gcs/presolvers/, which was fine
at two of them and stops being fine as the Cumulative presolver family
(#547, #548, #549) lands. Give each one its own directory, exactly as
every constraint has since #393:

    gcs/presolvers/auto_table.hh                     umbrella
    gcs/presolvers/auto_table/auto_table.{hh,cc}
    gcs/presolvers/difference_logic.hh               umbrella
    gcs/presolvers/difference_logic/difference_logic.{hh,cc}
    gcs/presolvers/difference_logic/difference_logic_test.cc

The umbrella keeps the bare include guard and the public include path:
<gcs/presolvers/auto_table.hh> and <gcs/presolvers/difference_logic.hh>
are what the examples, solve_test and the frontends include, and none of
them change. The class headers take the directory-qualified guard, the
implementation .cc files include the class header directly, and the test
keeps including the umbrella --- the same split constraints use.

Pure movement. Everything outside the renames is mechanical: three paths
in gcs/CMakeLists.txt, two include lines, two include guards, and three
comments and error messages that name the source file. The message in
difference_logic_test.cc is rewrapped only because the longer path no
longer fits its 80-column layout.

Evidence, with GCS_PRESERVE_PROOF_FILES=1 and a fixed --seed=12345 (the
presolver test randomises its branching, so an unseeded run is not
comparable): difference_logic_presolver_test in `proofs` and `opb` modes,
difference_chain --variant presolved --donor comparison in both modes,
and the auto_table and shikaku examples produce 1096 proof artefacts
between them --- 274 .opb, 274 .pbp, 274 .scp, 274 .varmap --- and every
one is byte-identical before and after. The two runs' stdout differs only
in wall-clock timing fields.

Issue #589.
